+++ src/GridCal/Gui/Main/ConsoleLog.py
@@ -3,7 +3,7 @@
 # Form implementation generated from reading ui file 'ConsoleLog.ui',
 # licensing of 'ConsoleLog.ui' applies.
 #
-# Created: Thu Dec  5 10:28:41 2019
+# Created: Thu Dec  5 22:15:14 2019
 #      by: pyside2-uic  running on PySide2 5.13.0
 #
 # WARNING! All changes made in this file will be lost!